Label Cell created
Deleted headerLabel from discoverCell and configured size and cell at HomeController
