--- AlwayGo/Cells/DiscoverCell/DiscoverCollectionCell.swift
@@ -1,83 +1,0 @@
-//
-//  DiscoverCollectionCell.swift
-//  AlwayGo
-//
-//  Created by Farid Rustamov on 09.03.25.
-//
-
-import UIKit
-
-class DiscoverCollectionCell: UICollectionViewCell {
-    
-    private lazy var titleLabel: UILabel = {
-        let label = UILabel()
-        label.font = UIFont(name: "SFPRODISPLAYMEDIUM", size: 20)
-        label.textColor = .label
-        label.text = "DISCOVER"
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    private lazy var collection: UICollectionView = {
-        var layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .vertical
-        layout.sectionInset = .init(top: 16, left: 16, bottom: 24, right: 16)
-        layout.minimumInteritemSpacing = 12
-        var collection = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collection.backgroundColor = .systemBackground
-        collection.showsHorizontalScrollIndicator = false
-        collection.translatesAutoresizingMaskIntoConstraints = false
-        collection.register(DiscoverCell.self, forCellWithReuseIdentifier: "\(DiscoverCell.self)")
-        return collection
-    }()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        
-        configureUI()
-        configureConstraints()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func configureUI() {
-        collection.backgroundColor = .systemBackground
-        collection.delegate = self
-        collection.dataSource = self
-    }
-    
-    private func configureConstraints() {
-        contentView.addSubview(titleLabel)
-        contentView.addSubview(collection)
-        
-        NSLayoutConstraint.activate([
-            titleLabel.topAnchor.constraint(equalTo: contentView.topAnchor),
-            titleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-            
-            collection.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
-            collection.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 8),
-            collection.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            collection.trailingAnchor.constraint(equalTo: contentView.trailingAnchor)
-        ])
-    }
-}
-
-extension DiscoverCollectionCell: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        10
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "\(DiscoverCell.self)", for: indexPath) as! DiscoverCell
-        
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        .init(width: 370, height: 270)
-    }
-    
-}
